+++ Tests/MMIOFileCheckTests/Tests/TestRegisterModifyProjectedMismatchedWidthIsStaticTrap.swift
@@ -11,28 +11,28 @@
 
 import MMIO
 
-@Register(bitWidth: 8)
+@RegisterDescriptor(bitWidth: 8)
 struct R8 {
   @ReadWrite(bits: 0..<2, as: Bool.self)
   var lo: LO
 }
 let r8 = Register<R8>(unsafeAddress: 0x1000)
 
-@Register(bitWidth: 16)
+@RegisterDescriptor(bitWidth: 16)
 struct R16 {
   @ReadWrite(bits: 0..<2, as: Bool.self)
   var lo: LO
 }
 let r16 = Register<R16>(unsafeAddress: 0x1000)
 
-@Register(bitWidth: 32)
+@RegisterDescriptor(bitWidth: 32)
 struct R32 {
   @ReadWrite(bits: 0..<2, as: Bool.self)
   var lo: LO
 }
 let r32 = Register<R32>(unsafeAddress: 0x1000)
 
-@Register(bitWidth: 64)
+@RegisterDescriptor(bitWidth: 64)
 struct R64 {
   @ReadWrite(bits: 0..<2, as: Bool.self)
   var lo: LO